requirements.txt
pandas
matplotlib
pyserial